solved integration error

+++ models/loyaltyProgramModel.js
@@ -26,17 +26,24 @@ const addLoyaltyProgram = async (loyalty_Programs_Data) => {
   const updateLoyaltyProgram = async (loyalty_Programs_updateData) =>{
     const {program_id,  name, description, points_per_dollar } = loyalty_Programs_updateData;
     try {
-        const query = `
-            UPDATE LoyaltyPrograms SET
-             name =COALESCE($1, name), 
-             description = COALESCE($2, description), 
-             points_per_dollar =COALESCE($3, points_per_dollar)
-             WHERE 
-             program_id = $4
-        `;
-        const values = [name, description, points_per_dollar, program_id];
+      if(name){
+        const query = 'UPDATE LoyaltyPrograms SET name = $1 WHERE program_id = $2';
+        const values = [name, program_id];
         await pool.query(query, values);
         return { success: true };
+      }
+      else if(description){
+        const query = 'UPDATE LoyaltyPrograms SET description = $1 WHERE program_id = $2';
+        const values = [description, program_id];
+        await pool.query(query, values);
+        return { success: true };
+      }else if(points_per_dollar){
+        const query = 'UPDATE LoyaltyPrograms SET points_per_dollar = $1 WHERE program_id = $2';
+        const values = [points_per_dollar, program_id];
+        await pool.query(query, values);
+        return { success: true };
+      }
+      
       } catch (error) {
         console.error('Error updateing loyalty program data:', error.message);
         return { success: false, error: 'Internal Server Error' };
@@ -44,9 +51,19 @@ const addLoyaltyProgram = async (loyalty_Programs_Data) => {
     
   }
   const deleteLoyaltyProgram = async(program_id) =>{
-
-    await pool.query('DELETE FROM LoyaltyPrograms WHERE program_id = $1', [program_id]);
+  
+   try{
+    if (program_id) {   
+      await pool.query('DELETE FROM LoyaltyPrograms WHERE program_id = $1', [program_id]);
+      return { success: true };
+  } else {
+    console.log('Program ID is required.');
   }
+  }catch (error) {
+    console.error('Error deleting Loyalty Programs:', error.message);
+    return { success: false, error: 'Internal Server Error' };
+  }
+}
 
 
 module.exports = {
